requirements.txt
pathfinding==1.0.4
python-socketio==5.8.0
aiohttp==3.8.5
pydantic==2.2.1
torch==2.0.1
transformers==4.33.1
sentence_transformers==2.2.2
accelerate==0.22.0
qdrant-client==1.5.0
langchain==0.0.283